requirements.txt
altgraph==0.17.4
appdirs==1.4.4
asn1==3.1.0
ConfigUpdater==3.2
enum-compat==0.0.3
grpcio==1.74.0
macholib==1.16.3
numpy==2.3.2
mido==1.3.3
packaging==25.0
protobuf==5.26.1
psutil==7.0.0
py-ptsl==500.1.1
py4j==0.10.9.9
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
Pypubsub==4.0.3
python-osc==1.9.3
python-rtmidi==1.5.8
six==1.17.0
wxPython==4.2.3
# Keep this version of setuptools, needed for py2app currently
setuptools==70.3.0